Add Nameable class
