--- Student.rb
@@ -1,12 +1,0 @@
-require './Person.rb'
-
-class Student < Person
-  def initialize(classroom)
-    super
-    @classroom = classroom
-  end
-
-  def play_hooky
-    "¯\(ツ)/¯"
-  end
-end
