--- Teacher.rb
@@ -1,11 +1,0 @@
-require './Person.rb'
-
-class Teacher < Person
-  def initialize(specialization)
-    super
-    @specialization = specialization
-  end
-
-  def can_use_services?
-    true
-  end
