+++ nameable.rb
@@ -1,0 +1,5 @@
+class Nameable
+  def correct_name
+    NotImplementedError.new('You must implement the correct_name method')
+  end
+end
